Html scoreboard
I made a connection between my SQLite database and html python code and made a html table showing all of the SQLite database in html form. I also made a system to automatically insert "John Doe" or "Jane Doe" as a defualt name if the user didnt write a name.